FIRST CRUD done. Workers. Also changed Long to Integer for DB Compatibility

+++ Java/vilchezmax/src/main/java/db/dao/IBaseDAO.java
@@ -2,7 +2,10 @@ package db.dao;
 
 public interface IBaseDAO<T> {
     void create(T object);
-    T getById(Long id);
+
+    T getById(Integer id);
+
     void update(T object);
-    void remove(Long id);
+
+    void remove(Integer id);
 }
